+++ packages/stripe_js/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 6.4.0
+- Use latest platform interface
+
+## 6.3.0
+**Features**
+- Add P24 support on Stripe web
+
+**Fixes**
+- Compliance with latest Dart analyzer
+- Prevent focus loop in Payment element for the web
+
 ## 6.2.1
 **Features**
 - Add support for any kind of payment method in setup intent with [confirmSetup]